Initial player logging implementation
Added a feature for plugin to send notification of player joining or
leaving the server. These changes might not be final as formatting might
change or extra functionality might need to be added.

I also had the liberty to clean up some parts of the code: removed
MsgTypes enum, used existing logger.

MsgTypes enum was replaced with just premade static messages for when
sever is going online/offline. As for other messages I just created
message objects in the places where they are required to be sent instead
of having static objects.

As for logger, instead of storing new/existing logger for the plugin, I
found that JavaPlugin already exposes it as public field, so started
using it. Fair WARNING as this could change later in the future.