Make event handling send Discord async messages
Makes CustomEventListener send discord notifications asynchronously.
Previously it would send them on the same thread as the event is called
on, which counts as main thread due to how minecraft is designed. That
behaviour in some rare cases could cause a small lag on the server as
it holds back the main thread. With this commit that should no longer be
the case and should be simpler solution than implementing
multi-threading.